Tools: Fixed error in podofonoc and warnings in all tools

+++ tools/podofocountpages/countpages.cpp
@@ -76,7 +76,7 @@ int main(int argc, char* argv[])
     }
     catch (PdfError& e)
     {
-        fprintf(stderr, "Error: An error %i ocurred during counting pages in the pdf file.\n", e.GetCode());
+        fprintf(stderr, "Error: An error %i ocurred during counting pages in the pdf file.\n", (int)e.GetCode());
         e.PrintErrorMsg();
         return (int)e.GetCode();
     }
