+++ tools/podofocrop/podofocrop.cpp
@@ -210,7 +210,7 @@ int main(int argc, char* argv[])
     }
     catch (PdfError& e)
     {
-        fprintf(stderr, "Error: An error %i ocurred during croppping pages in the pdf file.\n", e.GetCode());
+        fprintf(stderr, "Error: An error %i ocurred during croppping pages in the pdf file.\n", (int)e.GetCode());
         e.PrintErrorMsg();
         return (int)e.GetCode();
     }
