+++ tools/podofoimgextract/podofoimgextract.cpp
@@ -38,7 +38,7 @@ int main(int argc, char* argv[])
     }
     catch (PdfError& e)
     {
-        fprintf(stderr, "Error: An error %i ocurred during processing the pdf file.\n", e.GetCode());
+        fprintf(stderr, "Error: An error %i ocurred during processing the pdf file.\n", (int)e.GetCode());
         e.PrintErrorMsg();
         return (int)e.GetCode();
     }
